parser/requirements.txt
web3>=6.0.0
python-dotenv>=1.0.0
fastapi>=0.68.0
uvicorn>=0.15.0
requests>=2.26.0
